Rename headers and footers
